+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtLintRunner.kt
@@ -1,0 +1,8 @@
+package org.jlleitschuh.gradle.ktlint
+
+import java.io.File
+import org.gradle.api.file.FileCollection
+
+interface KtLintRunner {
+    fun lint(ktlintClasspath: FileCollection, ktlintVersion: String, ignoreFailures: Boolean, ktlintArgsFile: File)
+}