+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/KtlintFormatTask.kt
@@ -3,6 +3,7 @@ package org.jlleitschuh.gradle.ktlint
 import java.io.PrintWriter
 import javax.inject.Inject
 import org.gradle.api.file.FileTree
+import org.gradle.api.file.ProjectLayout
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.tasks.CacheableTask
 import org.gradle.api.tasks.OutputFiles
@@ -10,8 +11,9 @@ import org.gradle.api.tasks.TaskAction
 
 @CacheableTask
 open class KtlintFormatTask @Inject constructor(
-    objectFactory: ObjectFactory
-) : BaseKtlintCheckTask(objectFactory) {
+    objectFactory: ObjectFactory,
+    projectLayout: ProjectLayout
+) : BaseKtlintCheckTask(objectFactory, projectLayout) {
     override fun additionalConfig(): (PrintWriter) -> Unit = {
         it.println("-F")
     }